requirements.txt
requests>=2.25.1
Pillow>=8.0.0
numpy>=1.19.0
scipy>=1.5.0
rembg>=2.0.0
librosa>=0.9.0
soundfile>=0.10.0
